Moving Sm4shSound to Sm4shMusic, few fixes
sm4shexplorer 0.06.2:
Add a popup before building the mod if the release or debug folder is
not empty, letting you choose if you want to keep going or not. It will
now clean the export/release (or debug) folder before beginning to
export.

sm4shmusic 0.04:
- Original/Remix and other labels fixed [TESTED]
- Load from saved xml -should- be fixed
- MyMusic Index + 1 fixed [TESTED]
- EUR game region selection -should- be fixed. My bad EUR code is 4 and
not 3 (have no idea what 3 is then)
- Whenever you select a pre existing blank file, it will ask you if you
want to change the properties to make it visible in Sound Test. [TESTED]
- One rare crash fixed while launching Sm4shMusic after a compilation