Sm4shExplorer 0.06.4b and Sm4shMusic 0.6b
Fix for libvgmstream: Some people claim they didnt have the bug after
this version, so hopefully that'll work for everyone.
Fix for sound desync that some people who removed a song (or blank) were
having. Sadly I don't really see a way around starting from scratch,
unless you know how to edit the saved xml, but that's seriously gonna
take longer.
Unsigned values have been changed to signed (4294967295 will be shown as
-1 now), please understand that it breaks the deserialization of any
saved xml config.
Rarity changed to 100 by default
Unk5 is now known as Play Delay thanks to @Pib
BLANK files are now automatically removed when loading the list. This is
to avoid a lot of confusion. Instead when you click Add, Sm4shMusic will
now automatically fill the gaps in IDs (like in a config from scratch,
the first ID to be created will be 15)
MSBT variables names are now sync to the ID of the song, base 36 (so
like ID 10 will be SOUNDA, ID 11 will be SOUNDB, etc). It means that the
xml doesn't have to hold the information about the label anymore, making
it smaller.